+++ clock/time.go
@@ -1,0 +1,8 @@
+package clock
+
+import "time"
+
+// DateTimeNow returns the current datetime in standard format
+func DateTimeNow() DateTime {
+	return StandardFormat(time.Now())
+}