+++ list/list.go
@@ -1,3 +1,4 @@
+// Package list contains list functions
 package list
 
 import (