+++ my/my_test.go
@@ -1,0 +1,13 @@
+package my
+
+import "testing"
+
+func TestStatusCodes(t *testing.T) {
+	// TODO: OK200, OK201
+	// TODO: Err401, 403, 404, 429, 500
+}
+
+func TestInstance(t *testing.T) {
+	// TODO: NewInstance
+	// TODO: Instance.AddConnection
+}