+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.3.2 - Model Package 
+  * Commit: 
+## v0.3.1 - My Package 
+  * Commit: 2026-03-31 15:26
+  * Request type
+  * Request methods
+  * Clock format functions
 ## v0.3.0 - Rename Package
   * Commit: 2026-03-28 21:56
   * Rename package to github.com/zeroibot/pack